# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Vertice Coach API
  description: API for managing trainers, students, and training plans
  version: 1.0.0

tags:
  - name: Trainers
    description: Trainer management

paths:
  /api/trainers:
    get:
      tags: [Trainers]
      operationId: listTrainers
      summary: List all trainers
      responses:
        '200':
          description: List of trainers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainerResponse'

    post:
      tags: [Trainers]
      operationId: createTrainer
      summary: Create a new trainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerRequest'
      responses:
        '201':
          description: Trainer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerResponse'
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Email already in use
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /api/trainers/{id}:
    get:
      tags: [Trainers]
      operationId: getTrainer
      summary: Get a trainer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Trainer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerResponse'
        '404':
          description: Trainer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    put:
      tags: [Trainers]
      operationId: updateTrainer
      summary: Update a trainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerRequest'
      responses:
        '200':
          description: Trainer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerResponse'
        '404':
          description: Trainer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Email already in use by another trainer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      tags: [Trainers]
      operationId: deleteTrainer
      summary: Delete a trainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Trainer deleted
        '404':
          description: Trainer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  schemas:
    TrainerRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1

    TrainerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string